Lagt til blacklist for deltakerliste tombstone-meldinger

+++ amt-aktivitetskort-publisher/src/main/kotlin/no/nav/amt/aktivitetskort/service/KafkaConsumerService.kt
@@ -18,6 +18,7 @@ import no.nav.amt.aktivitetskort.utils.RepositoryResult
 import no.nav.amt.lib.models.deltaker.DeltakerKafkaPayload
 import no.nav.amt.lib.models.deltaker.DeltakerStatus
 import no.nav.amt.lib.models.deltakerliste.kafka.GjennomforingV2KafkaPayload
+import no.nav.amt.lib.models.deltakerliste.kafka.GjennomforingV2KafkaPayload.Companion.deltakerlisteTombstoneBlacklist
 import no.nav.amt.lib.utils.unleash.CommonUnleashToggle
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
@@ -88,7 +89,9 @@ class KafkaConsumerService(
         value: String?,
     ) {
         if (value == null) {
-            deltakerlisteRepository.delete(id)
+            if (id !in deltakerlisteTombstoneBlacklist) {
+                deltakerlisteRepository.delete(id)
+            }
             return
         }
